requirements.txt
django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.1